Refactoring code with style

+++ src/main/java/Exceptions/ArthiExcpetionCheck.java
@@ -7,7 +7,7 @@ package Exceptions;
 public class ArthiExcpetionCheck {
     public static void main(String[] args) {
         try {
-            int a = (int) 10 / 0;
+            int a = 10 / 0;
         } catch (Exception e) {
             e.printStackTrace();
         }
